let's hope this is the final commit-push of the project

+++ test/TEST.py
@@ -34,9 +34,6 @@ def clean_dir(is_interrupt,args):
     
 if __name__ == '__main__':
     try:
-        print('Hi! You are now testing our model! Relax, it will take a while...')
-        device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-
         model_choices = [
             'resnet50' , 'resnet50cbamcontrastive' , 
             'resnet152', 'resnet152cbam',
@@ -57,7 +54,13 @@ if __name__ == '__main__':
             action='store_true',
             help="Keep the test directory after the script finishes."
 )
+        parser.add_argument(
+            '--download_test_set',
+            action = 'store_true',
+            help = 'Download the test set (useful if testing on a local machine)'
+        )
         args = parser.parse_args()
+        print('Hi! You are now testing our model! Relax, it will take a while...')
         if not args.model in model_choices:
             print(f'Invalid model argument: {args.model}')
             print(f'Try one of {model_choices}')
@@ -67,7 +70,7 @@ if __name__ == '__main__':
             'resnet50':                 'https://drive.google.com/uc?id=150nfmRGFLTWo8uQ8W1t6cg73sZotOpXK',
             'resnet50cbamcontrastive':  'https://drive.google.com/uc?id=1JkFiuDOkt1Wq3nZAk7dr1XyvCIKJetkG',
             'resnet152':                'https://drive.google.com/uc?id=1HL--zu1VRlUEcc0bnuBiglXMDASSe2VY',
-            'resnet152cbamcontrastive':            'https://drive.google.com/uc?id=1pfnvDuPJ_oAUH9TawN-P8olULDaprEb9',
+            'resnet152cbam':            'https://drive.google.com/uc?id=1pfnvDuPJ_oAUH9TawN-P8olULDaprEb9',
             'vit':                      'https://drive.google.com/uc?id=19c2-zr7I9aAhAce5W0Tfd9MTpE9aS6_c',
             'wavenet':                  'https://drive.google.com/uc?id=1MuSdfqPd-yOR_ykk1ure7NUH3zTzGHLW', 
             'ensemble':                 'https://drive.google.com/uc?id=1dzsHe-BNYUIWzCzJN8Ktskmakh318TfL'
@@ -86,7 +89,7 @@ if __name__ == '__main__':
         torch.cuda.manual_seed_all(42)
         torch.backends.cudnn.enabled = False
 
-        model = nn.DataParallel(models_models_dict[args.model]).to(device)
+        model = nn.DataParallel(models_models_dict[args.model]).cuda()
         
         script_dir = os.path.dirname(os.path.abspath(__file__))
         os.chdir(script_dir)
@@ -97,21 +100,25 @@ if __name__ == '__main__':
             )
 
         weights_filename = [filename for filename in os.listdir() if 'pth' in filename][0]
-        model.load_state_dict(torch.load(weights_filename,weights_only=False,map_location=torch.device('cpu'))['model'])
+        model.load_state_dict(torch.load(weights_filename,weights_only=False)['model'])
         os.remove(weights_filename)
-        
         print('Weights file successfully loaded!')
-        print('Downloading the dataset')
-        gdown.download(
-                url='https://drive.google.com/uc?id=19nrUNb4U3PCgCDTUGFYNPlK1ZHZwI61S'
-            )
-        print('Extracting the dataset...')
-        with zipfile.ZipFile('test.zip') as zip_ref:
-            zip_ref.extractall()
-        os.remove('test.zip')
+        
+        test_dataset = DirectorySequentialDataset(dir='/work/cvcs2024/VisionWise/test')
+        if args.download_test_set:
+            
+            print('Downloading the dataset')
+            gdown.download(
+                    url='https://drive.google.com/uc?id=19nrUNb4U3PCgCDTUGFYNPlK1ZHZwI61S'
+                )
+            print('Extracting the dataset...')
+            with zipfile.ZipFile('test.zip') as zip_ref:
+                zip_ref.extractall()
+            os.remove('test.zip')
+            test_dataset = DirectorySequentialDataset(dir='test')
         #torch.use_deterministic_algorithms(True)
         #dataset and dataloader for testing
-        test_dataset = DirectorySequentialDataset(dir='test')
+
         test_dataloader = TransformDataLoader(
             cropping_mode=RandomTransform.GLOBAL_CROP,
             dataset=test_dataset,
@@ -124,7 +131,9 @@ if __name__ == '__main__':
 
         model.eval()
         with torch.no_grad():
-            accuracy = 0.0
+            good_answers = 0.0
+            good_real_answers = 0.0
+            good_fake_answers = 0.0
             max_iter = 0
             print('Dataset Iteration')
             iter_n = 0
@@ -133,18 +142,17 @@ if __name__ == '__main__':
                 test_pred = torch.sigmoid(model(test_images))
                 test_pred = torch.round(test_pred)
                 max_iter += test_pred.shape[0]
-                good_answers = torch.sum(test_pred == test_labels)
-                good_real_answers = torch.sum(test_pred == 0 & test_labels == 0)
-                good_fake_answers = torch.sum(test_pred == 1 & test_labels == 1)
-                
+                good_answers += torch.sum(test_pred == test_labels)
+                good_real_answers += torch.sum(torch.logical_and(test_pred == 0, test_labels == 0))
+                good_fake_answers += torch.sum(torch.logical_and(test_pred == 1, test_labels == 1))
+
+            
                 iter_n +=100
             
             print('Accuracy is -->' + str(good_answers.item()*100/max_iter) + '%')
-            print('Real accuracy is -->' + str(good_real_answers*200/max_iter) + '%')
-            print('Fake accuracy is -->' + str(good_fake_answers*200/max_iter) + '%')
+            print('Real accuracy is -->' + str(good_real_answers.item()*200/max_iter) + '%')
+            print('Fake accuracy is -->' + str(good_fake_answers.item()*200/max_iter) + '%')
             
         clean_dir(is_interrupt=False,args=args)
     except KeyboardInterrupt:
         clean_dir(is_interrupt=True,args=args)
-        
-
